add web response to auth builder

+++ src/Services/Traits/Profile/Logout.php
@@ -3,6 +3,7 @@
 namespace TomatoPHP\TomatoCrm\Services\Traits\Profile;
 
 use TomatoPHP\TomatoCRM\Helpers\Response;
+use TomatoPHP\TomatoCrm\Services\Contracts\WebResponse;
 
 trait Logout
 {
@@ -10,13 +11,18 @@ trait Logout
      * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function logout(\Illuminate\Http\Request $request): \Illuminate\Http\JsonResponse
+    public function logout(\Illuminate\Http\Request $request, string $type = "api"): \Illuminate\Http\JsonResponse|WebResponse
     {
         auth($this->guard)->logout();
 
-        $user = $this->model::find($request->user()->id);
-        $user->tokens()->delete();
+        if($type === 'api'){
+            $user = $this->model::find($request->user()->id);
+            $user->tokens()->delete();
 
-        return Response::success("Logout Success");
+            return Response::success("Logout Success");
+        }
+        else {
+            return WebResponse::make(__('Logout Success'))->success();
+        }
     }
 }
